+++ test/decode-test.js
@@ -1,0 +1,4 @@
+// Run test suite.
+describe.skip('morjs.decode', function() {
+  // TODO: Complete
+});